Unify and deduplicate stream and regular file loading
