+++ src/api/ude/Nacist-soubor.ts
@@ -1,3 +1,4 @@
+import { pipeline, Readable } from 'stream'
 import { z } from 'zod'
 
 import { Ginis } from '../../ginis'
@@ -10,6 +11,7 @@ import {
   RequestParamOrder,
   RequestParamType,
 } from '../../utils/request-utils'
+import { XmlBase64DataStreamParser } from './Nacist-soubor-stream'
 
 const nacistSouborRequestProperties = ['Id-souboru'] as const
 
@@ -48,6 +50,10 @@ export const nacistSouborResponseSchema = z.object({
 export type UdeNacistSouborNacistSoubor = z.infer<typeof nacistSouborSchema>
 export type UdeNacistSouborResponse = z.infer<typeof nacistSouborResponseSchema>
 
+const requestName = 'Nacist-soubor'
+const requestNamespace = 'http://www.gordic.cz/svc/xrg-ude/v_1.0.0.0'
+const requestXrgNamespace = 'http://www.gordic.cz/xrg/ude/nacist-soubor/request/v_1.0.0.0'
+
 export async function nacistSoubor(
   this: Ginis,
   bodyObj: UdeNacistSouborRequest
@@ -55,20 +61,64 @@ export async function nacistSoubor(
   const url = this.config.urls.ude
   if (!url) throw new GinisError('GINIS SDK Error: Missing UDE url in GINIS config')
 
-  const requestName = 'Nacist-soubor'
-  const requestNamespace = 'http://www.gordic.cz/svc/xrg-ude/v_1.0.0.0'
-
   const response = await makeAxiosRequest<string>(
     createXmlRequestConfig(requestName, requestNamespace),
     url,
     createXmlRequestBody(this.config, {
       name: requestName,
       namespace: requestNamespace,
-      xrgNamespace: 'http://www.gordic.cz/xrg/ude/nacist-soubor/request/v_1.0.0.0',
+      xrgNamespace: requestXrgNamespace,
       paramsBodies: [bodyObj],
       paramOrders: nacistSouborParamOrders,
     }),
     this.config.debug
   )
   return await extractResponseJson(response.data, requestName, nacistSouborResponseSchema)
+}
+
+/**
+ * Streaming variant downloading files from GINIS UDE (Úřední deska) service.
+ *
+ * **Why this exists:** The non-streaming version loads the entire SOAP XML
+ * response into memory, parses it with xml2js, then base64-decodes the file.
+ * For large files (50-200+ MB) this causes memory spikes of several GB
+ * (the base64 string + parsed object + decoded buffer all coexist in memory).
+ *
+ * @param bodyObj - Request parameters, typically `{ 'Id-souboru': fileId }`.
+ * @returns `Readable` stream of decoded binary data.
+ * @throws {GinisError} on network errors, SOAP faults, or malformed responses.
+ */
+export async function nacistSouborStream(
+  this: Ginis,
+  bodyObj: UdeNacistSouborRequest
+): Promise<Readable> {
+  const url = this.config.urls.ude
+  if (!url) throw new GinisError('GINIS SDK Error: Missing UDE url in GINIS config')
+
+  const response = await makeAxiosRequest<Readable>(
+    { ...createXmlRequestConfig(requestName, requestNamespace), responseType: 'stream' },
+    url,
+    createXmlRequestBody(this.config, {
+      name: requestName,
+      namespace: requestNamespace,
+      xrgNamespace: requestXrgNamespace,
+      paramsBodies: [bodyObj],
+      paramOrders: nacistSouborParamOrders,
+    }),
+    this.config.debug
+  )
+
+  const parser = new XmlBase64DataStreamParser({
+    responseValidation: {
+      requestName,
+      responseSchema: nacistSouborResponseSchema,
+    },
+  })
+  pipeline(response.data, parser, (error) => {
+    if (error && !parser.destroyed) {
+      parser.destroy(error instanceof Error ? error : new Error(String(error)))
+    }
+  })
+
+  return parser as Readable
 }
