+++ src/api/ude/index.ts
@@ -1,6 +1,5 @@
 import { detailDokumentu } from './Detail-dokumentu'
-import { nacistSoubor } from './Nacist-soubor'
-import { nacistSouborStream } from './Nacist-soubor-stream'
+import { nacistSoubor, nacistSouborStream } from './Nacist-soubor'
 import { seznamDokumentu, seznamDokumentuFilterArchiv } from './Seznam-dokumentu'
 import { seznamKategorii } from './Seznam-kategorii'
